+++ LionHeart-iOS/LionHeart-iOS/Global/Extensions/Float+.swift
@@ -1,0 +1,18 @@
+//
+//  Float+.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/07/13.
+//
+
+import Foundation
+
+extension Float {
+    static var half: Self {
+        return 0.5
+    }
+    
+    static var full: Self {
+        return 1.0
+    }
+}
