+++ LionHeart-iOS/LionHeart-iOS/Scenes/Onboarding/ViewControllers/GetFatalNicknameViewController.swift
@@ -10,32 +10,6 @@ import UIKit
 
 import SnapKit
 
-enum OnboardingFatalNicknameTextFieldResultType {
-    case fatalNicknameTextFieldEmpty
-    case fatalNicknameTextFieldOver
-    case fatalNicknameTextFieldValid
-
-    var errorMessage: String {
-        switch self {
-        case .fatalNicknameTextFieldEmpty:
-            return "입력된 내용이 없습니다."
-        case .fatalNicknameTextFieldOver:
-            return "10자 이내로 입력해주세요."
-        case .fatalNicknameTextFieldValid:
-            return "정상입니다"
-        }
-    }
-    
-    var isHidden: Bool {
-        switch self {
-        case .fatalNicknameTextFieldEmpty, .fatalNicknameTextFieldOver:
-            return true
-        case .fatalNicknameTextFieldValid:
-            return false
-        }
-    }
-}
-
 protocol FatalNicknameCheckDelegate: AnyObject {
     func checkFatalNickname(resultType: OnboardingFatalNicknameTextFieldResultType)
     func sendFatalNickname(nickName: String)
@@ -75,7 +49,7 @@ final class GetFatalNicknameViewController: UIViewController {
         return label
     }()
     
-    private let fatalNickNameTextfield = OnboardingTextfield(textFieldType: .fatalNickname)
+    private let fatalNickNameTextfield = NHOnboardingTextfield(textFieldType: .fatalNickname)
     
 
     public override func viewDidLoad() {
