+++ LionHeart-iOS/LionHeart-iOS/Scenes/Onboarding/ViewControllers/GetPregnancyViewController.swift
@@ -10,32 +10,6 @@ import UIKit
 
 import SnapKit
 
-enum OnboardingPregnancyTextFieldResultType {
-    case pregnancyTextFieldEmpty
-    case pregnancyTextFieldValid
-    case pregnancyTextFieldOver
-    
-    var errorMessage: String {
-        switch self {
-        case .pregnancyTextFieldEmpty:
-            return "입력된 내용이 없습니다"
-        case .pregnancyTextFieldValid:
-            return "정상입니다"
-        case .pregnancyTextFieldOver:
-            return "1에서 40 사이의 숫자를 입력해주세요."
-        }
-    }
-    
-    var isHidden: Bool {
-        switch self {
-        case .pregnancyTextFieldEmpty, .pregnancyTextFieldOver:
-            return true
-        case .pregnancyTextFieldValid:
-            return false
-        }
-    }
-}
-
 protocol PregnancyCheckDelegate: AnyObject {
     func checkPregnancy(resultType: OnboardingPregnancyTextFieldResultType)
     func sendPregnancyContent(pregnancy: Int)
@@ -64,7 +38,7 @@ final class GetPregnancyViewController: UIViewController {
         return label
     }()
     
-    private let pregnancyTextfield = OnboardingTextfield(textFieldType: .pregancy)
+    private let pregnancyTextfield = NHOnboardingTextfield(textFieldType: .pregancy)
     
     private let textBox: UIView = {
         let view = UIView()
